+++ guillotina/utils/__init__.py
@@ -21,6 +21,7 @@ from .misc import apply_coroutine  # noqa; noqa
 from .misc import find_container  # noqa
 from .misc import get_current_container  # noqa
 from .misc import get_current_request  # noqa
+from .misc import get_current_db  # noqa
 from .misc import get_current_transaction  # noqa
 from .misc import get_random_string  # noqa
 from .misc import get_registry  # noqa